don't write out ffi.rs to src dir (#30)
This file shows up as untracked, and therefore gets auto-added by systems like Jujutsu. It could be added to `.gitignore`, but the easier (and recommended) thing to do is to write this file out to the Cargo-defined `OUT_DIR`.

Ctrl-clicking to `ffi.rs` in rust-analyzer continues to work as usual.